+++ examples/resources/ispconfig_web_database/resource.tf
@@ -2,6 +2,6 @@ resource "ispconfig_web_database" "example" {
   database_name    = "myapp_db"
   parent_domain_id = 1
   type             = "mysql"
-  active           = "y"
+  active           = true
   quota            = 500
 }